Parameterize bottom sheet visual regression tests
